pkg/fms: Handle flash-device from the web UI
