+++ pkg/fms/system/tpl/sudoers.conf.tpl
@@ -5,7 +5,7 @@
 # necessary to avoid a complex prompting scheme to pass the user
 # credentials inwards.
 _gizmo ALL=(root) NOPASSWD:/usr/bin/tzupdate
-_gizmo ALL=(root) NOPASSWD:/usr/bin/setcap cap_net_raw\,cap_net_bind_service+ep /var/lib/gizmo/bin/netinstall-cli
+_gizmo ALL=(root) NOPASSWD:/usr/bin/setcap cap_net_admin\,cap_net_raw\,cap_net_bind_service+ep /var/lib/gizmo/bin/netinstall-cli
 
 # This allows the 'admin' user to transparently gain authority, as
 # most users of the Gizmo platform are not expected to be seasoned
