+++ pkg/fms/ui/p2/fragments/nav.p2
@@ -14,6 +14,7 @@
           <a class="nav-item" href="/ui/admin/setup/integrations">Integrations</a>
           <a class="nav-item" href="/ui/admin/setup/net-wifi">WiFi Settings</a>
           <a class="nav-item" href="/ui/admin/setup/net-advanced">Advanced Network</a>
+          <a class="nav-item" href="/ui/admin/setup/flash-device">Flash Device</a>
         </div>
       </div>
       <div class="nav-container">
